requirements.txt
scikit-learn
dash
pandas
numpy
plotly >= 5.4.0
pickle
mysql-connector-python
paho-mqtt